+++ src/api/routes/register_route.py
@@ -8,8 +8,8 @@ from sqlalchemy.exc import IntegrityError
 from werkzeug.exceptions import InternalServerError
 
 from api.services.users_services import create_user, add_user_to_db
-from db.database import db
 from api.utils import handle_integrity_error
+from api.db.database import db
 
 register_bp = Blueprint("register", __name__)
 
@@ -22,19 +22,15 @@ def register_user():
         add_user_to_db(new_user)
 
         return render_template('register.html', msg="New user added to DB!"), 200
-
     except ValueError as e:
         # Handle validation errors.
         return render_template('register.html', msg=str(e)), 400
-
     except IntegrityError as e:
         db.session.rollback()
         handle_integrity_error(e)
-
     except Exception as e:
         # Handle any other exceptions and errors
         raise InternalServerError(f'Registration failed! Please try again later!, Error: {str(e)}')
-
     finally:
         db.session.close()
